+++ CHANGES.md
@@ -18,6 +18,7 @@ Change Log
 * Fixed `WallGeometry` bug that caused walls covering a short distance not to render [#2897](https://github.com/AnalyticalGraphicsInc/cesium/issues/2897)
 * Fixed `PolygonGeometry` clockwise winding order bug. 
 * Fixed `BillboardCollection` bounding sphere for billboards with a non-center vertical origin. [#2894](https://github.com/AnalyticalGraphicsInc/cesium/issues/2894)
+* The globe depth is now rendered during picking when `Scene.depthTestAgainstTerrain` is `true`.
 * Fixed extruded `RectangleGeometry` bug for small heights. [#2823](https://github.com/AnalyticalGraphicsInc/cesium/issues/2823)
 
 ### 1.11 - 2015-07-01